+++ src/engine/Graphics/ImageSource.ts
@@ -5,16 +5,30 @@ import { Logger } from '../Util/Log';
 import { ImageFiltering } from './Filtering';
 import { Future } from '../Util/Future';
 import { TextureLoader } from '../Graphics/Context/texture-loader';
+import { ImageWrapping } from './Wrapping';
 
 export interface ImageSourceOptions {
   filtering?: ImageFiltering;
+  wrapping?: ImageWrapConfiguration | ImageWrapping;
   bustCache?: boolean;
 }
+
+export interface ImageWrapConfiguration {
+  x: ImageWrapping;
+  y: ImageWrapping;
+}
+
+export const ImageSourceAttributeConstants = {
+  Filtering: 'filtering',
+  WrappingX: 'wrapping-x',
+  WrappingY: 'wrapping-y'
+} as const;
 
 export class ImageSource implements Loadable<HTMLImageElement> {
   private _logger = Logger.getInstance();
   private _resource: Resource<Blob>;
   public filtering: ImageFiltering;
+  public wrapping: ImageWrapConfiguration;
 
   /**
    * The original size of the source image in pixels
@@ -57,15 +71,40 @@ export class ImageSource implements Loadable<HTMLImageElement> {
    */
   public ready: Promise<HTMLImageElement> = this._readyFuture.promise;
 
+  public readonly path: string;
+
+  /**
+   * The path to the image, can also be a data url like 'data:image/'
+   * @param path {string} Path to the image resource relative from the HTML document hosting the game, or absolute
+   * @param options
+   */
+  constructor(path: string, options?: ImageSourceOptions);
   /**
    * The path to the image, can also be a data url like 'data:image/'
    * @param path {string} Path to the image resource relative from the HTML document hosting the game, or absolute
    * @param bustCache {boolean} Should excalibur add a cache busting querystring?
    * @param filtering {ImageFiltering} Optionally override the image filtering set by [[EngineOptions.antialiasing]]
    */
-  constructor(public readonly path: string, bustCache: boolean = false, filtering?: ImageFiltering) {
+  constructor(path: string, bustCache: boolean, filtering?: ImageFiltering);
+  constructor(path: string, bustCacheOrOptions: boolean | ImageSourceOptions, filtering?: ImageFiltering) {
+    this.path = path;
+    let bustCache = false;
+    let wrapping: ImageWrapConfiguration | ImageWrapping;
+    if (typeof bustCacheOrOptions === 'boolean') {
+      bustCache = bustCacheOrOptions;
+    } else {
+      ({ filtering, wrapping, bustCache } = {...bustCacheOrOptions});
+    }
     this._resource = new Resource(path, 'blob', bustCache);
-    this.filtering = filtering;
+    this.filtering = filtering ?? this.filtering;
+    if (typeof wrapping === 'string') {
+      this.wrapping = {
+        x: wrapping,
+        y: wrapping
+      };
+    } else {
+      this.wrapping = wrapping ?? this.wrapping;
+    }
     if (path.endsWith('.svg') || path.endsWith('.gif')) {
       this._logger.warn(`Image type is not fully supported, you may have mixed results ${path}. Fully supported: jpg, bmp, and png`);
     }
@@ -82,9 +121,29 @@ export class ImageSource implements Loadable<HTMLImageElement> {
     imageSource.data.setAttribute('data-original-src', 'image-element');
 
     if (options?.filtering) {
-      imageSource.data.setAttribute('filtering', options?.filtering);
+      imageSource.data.setAttribute(ImageSourceAttributeConstants.Filtering, options?.filtering);
     } else {
-      imageSource.data.setAttribute('filtering', ImageFiltering.Blended);
+      imageSource.data.setAttribute(ImageSourceAttributeConstants.Filtering, ImageFiltering.Blended);
+    }
+
+    if (options?.wrapping) {
+      let wrapping: ImageWrapConfiguration;
+      if (typeof options.wrapping === 'string') {
+        wrapping = {
+          x: options.wrapping,
+          y: options.wrapping
+        };
+      } else {
+        wrapping = {
+          x: options.wrapping.x,
+          y: options.wrapping.y
+        };
+      }
+      imageSource.data.setAttribute(ImageSourceAttributeConstants.WrappingX, wrapping.x);
+      imageSource.data.setAttribute(ImageSourceAttributeConstants.WrappingY, wrapping.y);
+    } else {
+      imageSource.data.setAttribute(ImageSourceAttributeConstants.WrappingX, ImageWrapping.Clamp);
+      imageSource.data.setAttribute(ImageSourceAttributeConstants.WrappingY, ImageWrapping.Clamp);
     }
 
     TextureLoader.checkImageSizeSupportedAndLog(image);
@@ -145,7 +204,9 @@ export class ImageSource implements Loadable<HTMLImageElement> {
       throw `Error loading ImageSource from path '${this.path}' with error [${error.message}]`;
     }
     // Do a bad thing to pass the filtering as an attribute
-    this.data.setAttribute('filtering', this.filtering);
+    this.data.setAttribute(ImageSourceAttributeConstants.Filtering, this.filtering);
+    this.data.setAttribute(ImageSourceAttributeConstants.WrappingX, this.wrapping?.x ?? ImageWrapping.Clamp);
+    this.data.setAttribute(ImageSourceAttributeConstants.WrappingY, this.wrapping?.y ?? ImageWrapping.Clamp);
 
     // todo emit complete
     this._readyFuture.resolve(this.data);